access (access v0.1.0)

Functions for accessing and altering elements within nested data structures

This is a translation of Elixir's Access module and associated functions from the Elixir Kernel module.

It is usually very verbose to access values within deeply nested structures in Erlang. For example, let's reach into a nested map:

  Data = #{foo => #{bar => baz}},
  #{foo := Bar} = Data,
  #{bar := Baz} = Bar,
  Baz.
  %%=> baz

access provides functions that operate on deeply nested structures using patterns that describe paths into the structure. The above example could be rewritten like so:

  access:get_in(Data, [foo, bar]).
  %%=> baz

Or to put a value into the structure:

  access:put_in(Data, [foo, bar], quiz).
  %%=> #{foo => #{bar => quiz}}

Access patterns are especially useful when nested data contains lists. We can filter the pattern to only access within certain lists on custom predicates with filter/1 or access all elements in a list with all/0.

  Data = #{elements => [#{a => b}, #{c => d}, #{a => e}, #{c => d}]}.
  access:get_in(Data, [elements, access:filter(fun(M) -> is_map_key(a, M) end), a]).
  %%=> [b,e]
  access:put_in(Data, [elements, access:all(), f], g).
  %%=> #{elements =>
  %%=>    [#{a => b,f => g},
  %%=>     #{c => d,f => g},
  %%=>     #{a => e,f => g},
  %%=>     #{c => d,f => g}]}

Types

access_fun/2

-type access_fun(Data, CurrentValue) :: get_fun(Data) | get_and_update_fun(Data, CurrentValue).

any_container/0

-type any_container() :: any().

container/0

-type container() :: [{atom(), any()}] | map().

get_and_update_fun/2

-type get_and_update_fun(Data, CurrentValue) ::
    fun((get_and_update, Data, fun((term()) -> term())) -> {CurrentValue, container()} | pop).

get_fun/1

-type get_fun(Data) :: fun((get, Data, fun((term()) -> term())) -> container()).

key/0

-type key() :: term().

nil_container/0

-type nil_container() :: nil.

pattern/0

-type pattern() :: key() | access_fun(any(), any()).

structure/0

-type structure() :: container() | any_container() | nil_container().

Functions

all()

-spec all() -> access_fun(Data :: list(), CurrentValue :: list()).

Returns a function that accessses all elements in a list

element(Index)

-spec element(non_neg_integer()) -> access_fun(Data :: tuple(), CurrentValue :: term()).

Returns a function that accesses the element at the given index in a tuple.

Examples

  Data = #{a => {b, b, b}},
  access:put_in(Data, [a, access:element(1)], e).
  %%=> #{a => {c, b, b}}

filter(Predicate)

-spec filter(fun((term()) -> boolean())) -> access_fun(Data :: list(), CurrentValue :: list()).

Returns a function that accesses elements in a list which return true in the given predicate

get_and_update_in(Data, Tail, Fun)

-spec get_and_update_in(structure(), pattern(), Fun) -> {CurrentValue, structure()}
                     when
                         Fun :: fun((term() | nil) -> {CurrentValue, NewValue | pop}),
                         CurrentValue :: term(),
                         NewValue :: term().

Gets the current value within a structure and returns a structure with that value updated

get_in(Data, T)

-spec get_in(structure(), pattern()) -> term().

Gets a value within a data structure

pop_in(Data, Keys)

-spec pop_in(container(), pattern()) -> term().

Pops a value from a nested structure

put_in(Data, Keys, Value)

-spec put_in(structure(), pattern(), term()) -> term().

Puts a value into a nested structure

update_in(Data, Keys, UpdateFun)

-spec update_in(structure(), pattern(), fun((term()) -> term())) -> term().

Updates a value within a nested structure